src/requirements.txt
blinker==1.9.0
cffi==1.17.1
click==8.2.0
colorama==0.4.6
cryptography==45.0.2
Flask==3.1.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-MySQLdb==2.0.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.2.2
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysqlclient==2.2.7
packaging==25.0
pillow==12.0.0
psycopg2-binary==2.9.10
pyarmor.cli.core==7.6.7
pycparser==2.22
PyMySQL==1.1.1
python-dotenv==1.1.0
SQLAlchemy==2.0.41
typing_extensions==4.13.2
Werkzeug==3.1.3
WTForms==3.2.1
Flask-SocketIO==5.0
python-socketio==5.9
eventlet==0.33